==> Lessons/06_Lab_GLMsMEGAN.Rmd <==
---
title: "6: Part 1 - Generalized Linear Models"
author: "Environmental Data Analytics | John Fay and Luana Lima | Developed by Kateri Salk"
date: "Fall 2022"
output: pdf_document
geometry: margin=2.54cm
editor_options: 
  chunk_output_type: console
---

## Objectives
1. Answer questions on M5/A5
2. Answer questions on M6 - GLMs
3. Practice more application GLM to real datasets


## Set up
```{r, message = FALSE}
library(tidyverse)
install.packages("agricolae")
library(agricolae)   

EPAair <- read.csv("./Data/Processed/EPAair_O3_PM25_NC1819_Processed.csv", stringsAsFactors = TRUE)
# Set date to date format
EPAair$Date <- as.Date(EPAair$Date, format = "%Y-%m-%d")

Litter <- read.csv("./Data/Processed/NEON_NIWO_Litter_mass_trap_Processed.csv", stringsAsFactors = TRUE)
# Set date to date format
Litter$collectDate <- as.Date(Litter$collectDate , format = "%Y-%m-%d")

# Set theme
mytheme <- theme_classic(base_size = 14) +
  theme(axis.text = element_text(color = "black"), 
        legend.position = "top")
theme_set(mytheme)
```



## Visualization and interpretation challenge

Create three plots, each with appropriately formatted axes and legends. Choose a non-default color palette.

1. geom_density of ozone divided by year (distinguish between years by adding transparency to the geom_density layer).
2. geom_boxplot of ozone divided by year. Add letters representing a significant difference between 2018 and 2019 (hint: stat_summary). 
3. geom_violin of ozone divided by year, with the 0.5 quantile marked as a horizontal line. Add letters representing a significant difference between 2018 and 2019. 

```{r}
#Exercise 1:
Q1<-ggplot(EPAair, aes(x = Ozone, fill = as.factor(Year)))+
  geom_density(alpha = 0.5)+
                 scale_fill_manual(values =c("darkgreen", "darkblue"))+
                 labs(x = "Ozone AQI value", y = "Density", fill = "")
print(Q1)

Q2<-ggplot(EPAair, aes(x= as.factor(Year), y= Ozone))+
  geom_boxplot()+
  stat_summary(geom = "text", fun = max, vjust= -1, size =4, label = c("b", "a"))+
                 labs(x = "", y ="Ozone AQI value")+
                 ylim(0, 150)
print(Q2)
Q3<-ggplot(EPAair, aes(x= as.factor(Year, )))  
```

## Linear Regression

Important components of the linear regression are the correlation and the R-squared value. The **correlation** is a number between -1 and 1, describing the relationship between the variables. Correlations close to -1 represent strong negative correlations, correlations close to zero represent weak correlations, and correlations close to 1 represent strong positive correlations. The **R-squared value** is the correlation squared, becoming a number between 0 and 1. The R-squared value describes the percent of variance accounted for by the explanatory variables. 

For the NTL-LTER dataset, can we predict PM2.5 from Ozone?

```{r}

#Exercise 2: Run a linear regression PM2.5 by Ozone. Find the p-value and R-squared value. 
Ex2Regression<-lm(data=EPAair, PM2.5 ~ Ozone)
summary(Ex2Regression)
#p-value: very low so reject the null that there's no relationship between PM2.5 and Ozone. Rsquared: 13% of the variation is explained by this model.

#Exercise 3: Build a scatterplot. Add a line and standard error for the linear regression. Add the regression equation to the plot

Ex3<-ggplot(EPAair, aes(x = Ozone, y = PM2.5))+
  geom_point()+
   geom_smooth(method="lm")+  #shows the line with standard error
  
print(Ex3)


```


## AIC to select variables

What other variables can we add to improve model?

```{r}

#Exercise 4: Build correlation plots and identify more possible explanatory variables to add to the regression.
library(corrplot)
PeterPaulnutrients <- 
  PeterPaul.chem.nutrients %>%
  select(tn_ug:po4) %>%
  na.omit()
PeterPaulCorr <- cor(PeterPaulnutrients) #creates correlation matrix)
corrplot(PeterPaulCorr, method = "ellipse")

EPAairOzoneandPM2.5<-
  EPAair %>% 
  select(meanLatitude:Ozone) %>% #you could also do this by creating a dataframe that is a subset of just meanLatitude:Ozone
  na.omit()
EPAairOzoneandPM2.5corr<-cor(EPAairOzoneandPM2.5)
corrplot(EPAairOzoneandPM2.5corr, method = "ellipse")
corrplot.mixed(EPAairOzoneandPM2.5corr, upper = "ellipse") #another way to visualize the corr plo

#Exercise 5: Choose a model by AIC in a Stepwise Algorithm. Do the results from AIC match the variables you selected on Exercise 5?



EPAOzoneAIC<-lm(data=EPAairOzoneandPM2.5, PM2.5 ~ Ozone + meanLatitude + meanLongitude + Month + Year)  #since we only have six variables, we can include them all in the AIC, however, if we had more variables we would want to make informed choices from the correlation plot and select only those variables with a strong relationship.

step(EPAOzoneAIC)
#two iterations. In the first, it's saying by removing year, the AIC will be lower, which improves the model.

#Exercise 6: Run another regression using the variables selected on Exercise 6. Compare r-squared value with the one from Exercise 2. 

OzoneBest<-lm(data = EPAairOzoneandPM2.5, PM2.5 ~ Ozone + meanLatitude + meanLongitude + Month)
summary(OzoneBest)

```



## Litter Exercise

```{r}
# Wrangle the data
Litter.Totals <- Litter %>%
  group_by(plotID, collectDate, nlcdClass) %>%
  summarise(dryMass = sum(dryMass))

# Format ANOVA as aov
Litter.Totals.anova <- aov(data = Litter.Totals, dryMass ~ plotID)
summary(Litter.Totals.anova)   #result was that the means are not the same, so run post-hoc test.

# Extract groupings for pairwise relationships
Litter.Totals.groups <- HSD.test(Litter.Totals.anova, "plotID", group = TRUE)
Litter.Totals.groups$groups

Litter.Totals <- Litter.Totals %>%
  mutate( treatgroups = Litter.Totals.groups$groups[plotID,2])

# Graph the results
Litter.Totals.plot <- ggplot(Litter.Totals, aes(x = plotID, y = dryMass)) +
  geom_boxplot() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  stat_summary(geom = "text", fun = max, vjust = -1, size = 3.5,
               label = c("abc", "ab", "abcd", "bcd", "bcd", "a", 
                         "d", "abcd", "cd", "d", "abcd", "abcd")) +
  labs(x = "Plot ID", y = "Dry Mass (g)") +
  ylim(0, 85)
print(Litter.Totals.plot)

```

```{r}
#Exercise 7: Improve the plot```
arrange(Litter.Totals, treatgroups)



print(Litter.Totals.plot)
